src: move slot finding functions from config_file.c to slot.c module
