+++ include/slot.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#include <glib.h>
+
+#include "checksum.h"
+
 typedef enum {
 	ST_UNKNOWN = 0,
 	ST_ACTIVE = 1,
@@ -55,3 +59,22 @@ typedef struct _RaucSlot {
 	RaucSlotStatus *status;
 } RaucSlot;
 
+/**
+ * Finds a slot given its device path.
+ *
+ * @param slots a GHashTable containing (gchar, RaucSlot) entries
+ * @param device the device path to search for
+ *
+ * @return a RaucSlot pointer or NULL
+ */
+RaucSlot *find_slot_by_device(GHashTable *slots, const gchar *device);
+
+/**
+ * Finds a slot given its bootname
+ *
+ * @param slots a GHashTable containing (gchar, RaucSlot) entries
+ * @param botname the bootname to search for
+ *
+ * @return a RaucSlot pointer or NULL
+ */
+RaucSlot *find_slot_by_bootname(GHashTable *slots, const gchar *bootname);
